move Configure... files to new subfolder and refactor how accessed
