+++ client/src/components/CampaignCreator/TriggerEvents/TransformData/ConfigureBasicQuarries.tsx
@@ -1,4 +1,4 @@
-import { NodePillarLists, TypeCampaignData } from '../../static_data_file_configs/PresetCampaignConfig';
+import { NodePillarLists, TypeCampaignData } from '../../../static_data_file_configs/PresetCampaignConfig';
 
 const ConfigureBasicQuarries = (campaignSettings: TypeCampaignData) => {
     return ['node_quarry_1', 'node_quarry_2', 'node_quarry_3', 'node_quarry_4'].reduce((currentQuarryObject, node) => {
