+++ client/src/components/CampaignCreator/TriggerEvents/TransformData/ConfigureBonusQuarries.tsx
@@ -1,5 +1,5 @@
-import { campaignCreatorData } from '../../static_data_file_configs/CampaignCreatorConfig';
-import { NodePillarLists, TypeCampaignData } from '../../static_data_file_configs/PresetCampaignConfig';
+import { campaignCreatorData } from '../../../static_data_file_configs/CampaignCreatorConfig';
+import { NodePillarLists, TypeCampaignData } from '../../../static_data_file_configs/PresetCampaignConfig';
 
 const ConfigureBonusQuarries = (campaignSettings: TypeCampaignData) => {
     return ['node_quarry_1', 'node_quarry_2', 'node_quarry_3', 'node_quarry_4'].reduce((currentQuarryObject, node) => {
