+++ client/src/components/CampaignCreator/TriggerEvents/TransformData/ConfigureNemesis.tsx
@@ -1,4 +1,4 @@
-import { NodePillarLists, TypeCampaignData } from '../../static_data_file_configs/PresetCampaignConfig';
+import { NodePillarLists, TypeCampaignData } from '../../../static_data_file_configs/PresetCampaignConfig';
 
 const ConfigureNemesis = (campaignSettings: TypeCampaignData) => {
     return ['node_nemesis_1', 'node_nemesis_2', 'node_nemesis_3'].reduce((currentNemesisObject, node) => {
