+++ client/src/components/CampaignCreator/TriggerEvents/TransformData/index.tsx
@@ -1,0 +1,17 @@
+import ConfigureBasicQuarries from './ConfigureBasicQuarries';
+import ConfigureBonusQuarries from './ConfigureBonusQuarries';
+import ConfigureGear from './ConfigureGear';
+import ConfigureMilestones from './ConfigureMilestones';
+import ConfigureNemesis from './ConfigureNemesis';
+import ConfigurePrinciple from './ConfigurePrinciple';
+import ConfigureResourceList from './ConfigureResourceList';
+
+export const TransformData = {
+    BasicQuarries: ConfigureBasicQuarries,
+    BonusQuarries: ConfigureBonusQuarries,
+    Gear: ConfigureGear,
+    Milestones: ConfigureMilestones,
+    Nemesis: ConfigureNemesis,
+    Principle: ConfigurePrinciple,
+    ResourceList: ConfigureResourceList,
+};
